Refactor WaitHelper
